Move default interner and tls debugging to chalk_integration
