+++ chalk-integration/Cargo.toml
@@ -16,7 +16,7 @@ salsa = "0.10.0"
 chalk-macros = { version = "0.10.1-dev", path = "../chalk-macros" }
 chalk-derive = { version = "0.10.1-dev", path = "../chalk-derive" }
 chalk-engine = { version = "0.10.1-dev", path = "../chalk-engine" }
-chalk-ir = { version = "0.10.1-dev", path = "../chalk-ir", features = ["default-interner"] }
+chalk-ir = { version = "0.10.1-dev", path = "../chalk-ir" }
 chalk-rust-ir = { version = "0.10.1-dev", path = "../chalk-rust-ir" }
 chalk-solve = { version = "0.10.1-dev", path = "../chalk-solve" }
 chalk-parse = { version = "0.10.1-dev", path = "../chalk-parse" }
