+++ chalk-integration/src/db.rs
@@ -1,10 +1,10 @@
 use crate::error::ChalkError;
+use crate::interner::ChalkIr;
 use crate::lowering::LowerGoal;
 use crate::program::Program;
 use crate::query::{Lowering, LoweringDatabase};
+use crate::tls;
 use chalk_engine::forest::SubstitutionResult;
-use chalk_ir::interner::ChalkIr;
-use chalk_ir::tls;
 use chalk_ir::AssocTypeId;
 use chalk_ir::Canonical;
 use chalk_ir::ConstrainedSubst;
